factored out blitting constraints check in a seperate function
the checking of the boundaries when blitting was incomplete
for the addblitToBuffer() function when compared with the
blitToBuffer() function. Factored out the redundant checking
procedure into a new function.

Should fix another bug leading to issue koreader/koreader-base#401 and probably
more cases for specifying boundary boxes.

Repo: base


Repo: base