reactivate FFI blitbuffer module This patch should fix koreader/koreader#274 and crashes reported at koreader/koreader#288
1. The shifted x coordinates bug(koreader/koreader#274) is caused by
non-integer x, y or w, h parameters passed to paintRect. It's fixed
in checkBounds method of blitbuffer.

2. The landscape crash reported at koreader/koreader/#288 is a
historied bug in the Lua-C-API of blitbuffer. The paintRoundedCorner
sometimes write over the boundery of blitbuffer and corrupts the heap.
Now this should be fixed in the FII module since the setPixel method
checks very carefully on each pixel write.