Introduce per-target CFLAGS.
This rids us of the legacy Kindle compat crap we don't need on the K5.
There's still a few things to check though:

* I'm using a 'perfect' K5 TC, so that works, but depending on the TC
used for Kindle builds on the buildbot, we may need to reintroduce
some compat flags.

* I'm mostly sure every Kobo we care about runs on an A8, but if a valid
target kobo device doesn't, fix this.

* I have no idea what we want to do for Android, so go with generic
armv6, and hope the TC defaults to sane stuff.

(This also introduces a couple of extra targets, some of them mostly for
my own use, so I don't have to monkey patch each time, and some to get
rid of legacy stuff on Kindles).

Repo: base


Repo: base