we don't render glyphs to 4bpp buffers anymore
Freetype returns 8bpp worth of "intensity" values. In the past, we
created a 4bpp blitbuffer for each glyph, specific to a foreground/
background color combination. We don't do this anymore and rather
use the data as returned by Freetype. We then use a special blit
method to use this grayscale data from Freetype as an alpha mask
for blitting a special color (what was the foreground color before)
to the target blitbuffer.

Repo: base


Repo: base