add wrappers for mupdf API that will handle exceptions
This will provide wrapper functions for mupdf functions that
might throw exceptions.

These wrapper functions will catch exceptions and instead return
a dedicated error value.

This error value could be NULL for functions that return pointers,
could be -1 for functions that return ints or could be anything else.

Note that in any case the functions will have a fz_context as their
first argument - indepedent of it is demanded by the mupdf API (in
most cases, it's the same for the mupdf API, but there are exceptions
to that rule).

As a convention, the wrapper functions start with "mupdf_" rather
than "fz_" or "pdf_" as per the original mupdf API.