Simplify the wait_for_marker logic
Only wait for the previous update when queuing FULL updates.
Since all REAGL updates are FULL, this is good enough.

Always use MXCFB_WAIT_FOR_SUBMISSION on REAGL-awaire Kindle devices.
This might prove to be overkill, in which case we could get rid of the
whole refresh() signature change, since a simple check for refreshtype
FULL would be enough :).