do not handle Android app events/commands in timing critical section
we handled some events directly in the looper parsing section. This
led to funky behaviour: code was only executed in some parts (weird,
very, very weird behaviour). But we don't need to handle those events
there, since we pass the event on anyway. So it is some other part of
the application that is now in charge to handle the events.

The timing critical part (or what seems to be it, as the NDK documentation
does not say any special things about this code area) now only enqueues
the event for later processing.

Repo: base



Repo: base