emulated screen flash on Emulator
This (re-)introduces a simulated "flash" (for now: invert, invert back)
of refresh sections with the Emulator. Since this is only useful when
debugging, it is off by default. It can be enabled by setting an
environment variable EMULATE_READER_FLASH. Set it to the number of
microseconds you want the flash to take. Suggested are values 200-700,
depending on how closely you want to look at refreshes :-)

Repo: base


Repo: base