Make the UI slightly snappier on REAGL Kindles.
By not waiting for UI markers. The WAIT_FOR_SUBMISSION should be good
enough.
That matches what's done on the Aura, as a bonus.

Speaking of, send a semi-random but still valid update_marker when we're
not tracking them, instead of 0, which is an invalid marker ID.

Repo: base



Repo: base