fix
fix txt format changes when it was opened before. ( crengine will get
rendered height and width from cache if it was opened before )
this problem is not so obvious when the border width is small.  I set
the width to 50px and the text had been wrongly placed when opened with
cache existed.

Repo: crengine



Repo: crengine