fix koreader/crengine#2908,#2386 cre crash
- koreader/crengine#2908
- in rare circumstances, image height being 0 (dividing result is less
than 1) causes crash
- koreader/crengine#2386 and maybe other weird epub crash
- there is a function called autoboxchildren(), which is to remove
empty nodes when openning a new book. autoboxchildren() will be called
again when the embedded styles is turned off. usually it will not cause
problem.
- if there is an empty <br/ class="block"> tag (css: .block { display
: block;}) in the epub and the book is large enough, it will crash in
unpredictable ways.
- with embedded style, the <br/> tag was stored as a block node, if
embedded style is off, it is treated as inline node. I suspect after the
cache file is saved it does not allow some node manipulation because
turning off emmbedded when the book is openning for first time will not
crash. as all this has no visual effect to the book, I believe it is
better to partially disable this funtion, ie. insert or move nodes.