Reintroduce base git-rev (koreader/koreader-base#540)
git-rev was moved to the frontend in 540add954fb9af2353ca0f09cb3751d27016ac35
because it's used as a versioning tool for the entirety of KOReader. However,
there are uses for keeping an easy to reference base git-rev as well.

It looks like the easiest way to ensure CircleCI cache is updated is to use a key.

https://circleci.com/docs/2.0/configuration-reference/#save_cache

In our case we can use `{{ checksum "filename" }}` to decide whether base has
changed and whether we should save a new base cache.

Presently the cache from early September is always reused and hence base will
always be compiled in the frontend CircleCI build.