Fix styles cache corruption
When a style is applied to a node with el->setStyle(newstyle), a hash of this
style is calculated and used as a key to access/re-use/retrieve this style.
If we modify such a style after that, we do corrupt the hash/style relation,
which may change some other elements' style.
And it will also prevent the use of the saved style cache when re-opening
a document from its cache, as then the validation of the loaded style hash
against the saved hash fails. A full rendering would then be done, losing
the benefit of the cache.
This fix allows cache to be re-used, and documents with cache to be
re-opened really faster.

Repo: crengine



Repo: crengine