Fix getPageDocumentRange() on some pages (koreader/crengine#105)
On some pages (eg: that ends with some padding between an
image on this page, and some text on next page), there may
be some area which is rendered "final" without any content,
thus holding no node. We would then get no xpointer for end
of page (and features that use it would fail, ie: word
selection, getCurrentPageLinks(), getPageText()).
But we can get a xpointer by looking a few pixels back the
height of page.