Stylesheets: adds support for !important (koreader/crengine#166)
Simple rules:
CSS are applied (to each node) mostly in the order they appear in the
(concatenated) stylesheets, inline styles last.

When meeting a property with !important: set it to node's style,
and set a flag this property is important

When meeting a property without !important: check first if an important
flag has previously been set to this node property, and if yes, ignore
it. If no, set it and override any previous value.

Repo: crengine