html documents: proper handling of <BR> tags
(for .html documents only - <BR> in epub were and are fine)
<BR> were previously converted to <P> because of a fix for "bad LIB.RU
books" being applied in any case. This could prevent the style of the
container to be applied on HTML sub-parts after a <BR>, or the style
of <P> (with text-indent) to be applied after a <BR>.
<BR> are now correctly considered to be a block element, and css
helps displaying multiple new lines when multiple <BR> (these CSS
updates fix multiple <BR> in EPUB too).

Also fix unrelated clang warning:
  "& has lower precedence than !=; != will be evaluated first"

Repo: crengine


Repo: crengine