epub.css cleanup 0: convert end of line (dos to unix)
Incremental cleanup of epub.css.
The aim is to remove classname based selectors, as it's the
embedded stylesheet job to provide declarations for classes.
We should keep the few ones involving display: property, to
be included for previously opened document depending on
gDOMVersionRequested.

Step0: convert from dos fileformat to unix