Rework styles loading, avoid nodes styles re-init in somes cases
Styles loading (from <HEAD><STYLES> and <HEAD><LINK REL="stylesheet">)
were done in different and inconsistent ways, resulting in
some of them not being applied, or not being stored in the DOM and
cache, so not re-applied on re-rendering (eg: embedded styles off
then back on) or on re-loading from cache.
Mostly noticable with HTML documents, where linked stylesheets
were not stored in the cache.

It was mostly fine with EPUBs, except that it worked because of some
stylesheets push()/pop() not done correctly, that triggered a full
styles re-init when an EPUB has a single HTML with either <HEAD>
or a secondary linked css file. This fixes that case, and avoid
the full styles re-init on such books, providing up to a
25% speed up in loading time.
(This full styles re-init still happens and is necessary when an
EPUB contains embedded fonts. Added a printf() so we'll know that,
and know when we may still have a bug with this.)

Background: while loading a document, styles are applied to nodes
during that loading phase (this is necessary mostly because of
"white-space: pre", so the right method to read text nodes content
is used), so it has not to be done again in the following render
phase (it is still re-done when we later change some rendering settings).
Before rendering, checkRenderContext ensures everything is coherent
with various hashes, and it ought to be coherent on first load.
Failing the push()/pop() resulted in the original external stylesheet
being no more the same as originally, and this was noted by some hash
check, triggerring an unnecessary full node styles re-init.

Repo: crengine



Repo: crengine