Fix support of style "white-space: pre"
"white-space: pre" in stylesheets were processed correctly, but
had no effect because it was applied too late (after the text
nodes has been loaded with a non-pre XML reading method, so
white spaces and newlines were no more there).
(It worked only on real <PRE> tags, because they have hardcoded
"white-space: pre" in fb2def.h.)
Also accounts for "white-space: pre" changes in the
nodeDisplayStyleHash, so we can let the user know that reloading
is needed for a correct rendering.