Adds ldomXPointer::getHtml() and ldomXRange::getHtml()
Also adds LVDocView::getDocumentFileStream() to get the content
of individual files in an EPUB.

Will allow viewing the HTML from some text selection (including
the stylesheets files used), and getting the HTML from the target
of some links to show them in a footnote popup.
These use an added versatile writeNodeEx() function with many flags
to tweak the produced HTML depending on needs.