Adds ldomXRange::getSegmentRects()
Returns the multiple segments (rectangle for each text line) that
a ldomXRange spans on the page.

Will be used to get the full area spanned by a link, and to
compute a more accurate distance from the gesture position (to
allow non-precise tap on links).
WIll also be used to get a more proper highlighting of selected
text than the current word by word one, which misses punctuation
on left and right edges.

Repo: crengine


Repo: crengine