Avoid wrap on a space before or after some specific chars
Avoid a '(' at end of line, or a ')' at start of line, or a
'?' at start of line with french texts (where some punctuation
requires a space before).
It's the work of the publisher to use a no-break-space in these
cases, but let crengine does a bit of this work he has not done.

Repo: crengine



Repo: crengine