Avoid uneeded re-rendering on load when with full status bar (koreader/crengine#238)
getPageHeaderHeight() was called before m_infoFont was set to
its definitive value. So, the second time it is called, we may
get a different value, which causes a difference in page height,
so a hash mismatch, and so a re-rendering.
Just move the dx/dy computation AFTER m_infoFont is set to
its correct and definitive value, and avoid any mismatch.

Repo: crengine