[build] Pull in thirdparty repos with --depth 50 (koreader/koreader-base#675)
Uses a fairly conservative value of 50 for now. Also implements a fallback to pull in the entire repo if that's not enough.

This will mainly help to speed up the nightly builds and the CI builds, but it will also reduce the space and time investment for potential hackers.

Submodules of thirdparty repos unfortunately won't be pulled in shallow yet.