Fix some issues when rendering text in constrained width
Issues mostly noticable in table cells:
- Really ignore leading spaces, so they don't cause a wrap
(and so a leading blank line) when the next word doesn't
fit in the available width.
- When a word exactly fits in the available width, don't
forget to check if the following char (even if it itself
doesn't fit) is a space, in which case it is a candidate
for wrap. Otherwise, such a word would be splitted by
the hyphenation rules (This also helps in normal text
when the last word fits exactly the line: it will avoid
hyphenating that last word, and reduce the amount of added
spacing between words because of that.)
- Fix some crash when some image gets resized to have
a zero width (can happen with some mix of CSS properties)

Repo: crengine


Repo: crengine