Page splitting: fix possible missing blocks
When consecutive lines or blocks with all page-break: avoid,
if they wouldn't fit on a page, some of them may be missing
in the computed pages (while still visible in scroll mode).
This was because StartPage() did set 'last=line', which
was wrong in the case we split on an earlier line (when
all current lines have pb avoid, we split on a past one
where page break was allowed we remembered (as 'next').
(Mostly noticable with tables with tall rows: currently,
a table with 3 rows has all pb avoid between rows).
This commit's real change is just the removal of 'last=line'
in StartPage().
Also moved the single 'last = line' that was at the end
of AddLine() to the proper logical place in each of the
if/else branches as it makes more sense there when reading.
Added many comments and debugging code (lvpagesplitter
may need a revisit.)

Repo: crengine



Repo: crengine