CSS: adds support for 'auto', ignore % for borders
- Adds support for named value 'auto' in some length'ed
properties (width: and margin:, will only be supported
for now when applied to a TABLE element).
- Remove support of units in % for border* properties
according to CSS specs (in some cases, when the container
width is not yet known, it couldn't be computed anyway).
- Fixed computing of border named values (thin, medium,
thick) that I forgot to *256 when I added support for
dpi-based css units (koreader/crengine#213).

Repo: crengine



Repo: crengine