Fix right border drawing position
It was too short by 1px since koreader/crengine#211 (bottom border needed koreader/crengine#211 fix,
and I added in koreader/crengine#211 the same kind of fix for the right border as it
looked needed - but the more obvious error, easier seen
on the right than on the bottom, was already workarounded by
some removed '+1' in the their drawing code.
Now, they are all similar.