Adds getRenderedWidths(): get node min/max node content width
Needed for proper sizing and positionning of some elements in
various contexts: will be used to properly size table cells,
but could be useful if we were to add support for 'float:'.

Estimate min and max content width of node when rendered,
the way it is mentionned in various CSS or HTML specs:
- maxWidth: width if it would be rendered on an infinite
  width area, breaking only on explicit <BR/>.
- minWidth: width with a break on all spaces (no hyphenation),
  so the width taken by the longest word or image.

This does mostly what renderBlockElement, renderFinalBlock and
lvtextfm.cpp do, but caring only about widths and horizontal
margin/border/padding and indent (with no width constraint, so no
line splitting and hyphenation - and we don't care about vertical
spacing and alignment).

Repo: crengine


Repo: crengine