CSS: Implement -cr-hint: footnote-inpage
This feature was already implemented in crengine, but could only be
witnessed with some specific FB2 documents.
Setting "-cr-hint: footnote-inpage" to each footnote block elements
will allow having them with EPUB documents.

The performance and memory usage penalty of gathering all internal
links for page splitting and footnotes inclusion seems really minimal
(tested with a big book full of footnotes), so this will now always
be done, even when there is not a single -cr-hint: footnote-inpage
(as we don't now that yet while the book is rendered).

Repo: crengine


Repo: crengine