In-page footnotes: fix ignored link in nested tags
"123" in <a href=><sup>123</sup></a> would not be flagged
as a link because it's not a direct child of the <A>, and
would be ignored in in-page footnotes.
Fix that by propagating a flag when rendering children
elements. The first non space-only text node will reset
that flag, as we're only interested in the first word
of a link being flagged.

Repo: crengine



Repo: crengine