cre:findText(): fix search getting stuck on some pages
We could get stuck on some page (not able to go see
next or previous pages that do contain the search
pattern) when this page (or its neighbours) starts
with some margin or empty elements that would not
resolve to a XPointer.
This ensures we always find a XPointer to work with
by moving a bit the provided y till we find one.

Repo: crengine



Repo: crengine