Harfbuzz: fix rendering when hyphenating on a ligature
When hyphenation happens on a ligature (i.e. with the word
"diffuse" being hyphenated "dif-fuse", with a font that
happens to have a ligature for "ff"), widths of each part
of the cut word would be wrong. We now re-measure the width
of each part, for correct positionning and spacing.