2-pages mode: tweak middle margin sizing
Fix previous code that was doubling the side margins to make
the middle margins (so, middle margin were 4x left margin).
Furthermore adjust this middle margin to be min 1.2em
and max only 1x left margin (or right margin if greater), which
seems to be a fine rule with any margin settings and font size.

Repo: crengine


Repo: crengine