Text rendering: fix words stuck when hyphenating on italic
We adjust the width of the last char on a line when it's in italic
to account for excessive right side bearing common with italic.
But we should not do that when this last char happens to be from
some hyphenated word, as it will be followed by an hyphen, so the
hyphen is the actual last char.
This would cause some errors in both this hyphenated word's
half-words widths, and some bad spacing around them.

Repo: crengine


Repo: crengine