Fix getPageDocumentRange() on some pages (again)
Similar to what f3976493 did for end of page, but
done now in a generic way for start and end, in page
and in scroll modes, as it solves some edge cases:

On some pages with some padding or margin at start
or end, there may be some area which is rendered
"final" without any content, thus holding no node.
We would then get no xpointer for start or end of
page (and features that use it would fail, ie: word
selection, getCurrentPageLinks(), getPageText()).
So try getting a xpointer by looking a few pixels
before or after.

Repo: crengine



Repo: crengine