[fix] DjVu gamma correction (koreader/koreader-base#921)
The existing implementation only worked in 8-bit mode. By using the `ddjvu_format_set_gamma()` function that comes with djvulibre it suddenly and attractively becomes Not Our Problem™.

I noticed this after doing a quick experiment with a custom power function `#ifdef __ANDROID_API__` like in https://stackoverflow.com/a/19488271/2470572

(For example, if Android define a custom power function, else `#define power(x, y) pow(x, y)`.)

I believe the proper way to go about it would probably be to ignore document-specific implementations and to do it in the blitbuffer, but this is quicker.

As a side effect, this fixes https://github.com/koreader/koreader/issues/3493.

Repo: base