Force a re-render when :last-child & other pseudoclasses are met
In the initial loading phase, when the HTML is parsed and the DOM
is being built, and styles computed (cause we know enough: the node
class, all the parents), when meeting the nodes as they are parsed,
the current node is... always the last child (which might be the only
thing we don't know yet...). So, some CSS pseudoclasses like:
    :last-child :last-of-type
    :nth-last-child() :nth-last-of-type()
    :only-child :only-of-type
might be matched on more nodes that they should.
So, when meeting such a pseudoclass, we flag that a re-render
is needed, which will be done on the fully made DOM, so they
are matched correctly.

Repo: crengine



Repo: crengine