Optimize background image drawing
Previous implementation would take many seconds to draw a background
image set on a block element (eg. BODY) if that element spans many
dozens of pages, as it would create a buffer that tall to finally
crop and paste only a small bit of it on the page buffer.
Optimize that by using crengine facilities.
Also, for BODY, try to have a fixed background when in page mode
to avoid ghosting on eInk (in scroll mode, keep it scrolling
with the elements).

Repo: crengine


Repo: crengine