Fonts: use Freetype embolden API for fake bold
Use Freetype embolden API instead of LVFontBoldTransform to
make fake bold (for fonts that do not provide a bold face).
This allows Harfbuzz to work with them (even if they won't
look as nice as if they came with a real bold font).
(LVFontBoldTransform is a wrapper that enlarge glyphs and
advances, so killing any Harfbuzz measurement and glyph
replacements.)
See comments for more details.