Fonts: handle RTL, rework Harfbuzz measuring and drawing
measureText(): accepts an added "hints" parameter,
thru which we can pass text direction and additional
hints for Harfbuzz (begin/end of paragraph for now).

For freetype and harbuzz light, draw chars in the
reverse order when direction is RTL: this allows
individual RTL words to be drawn correctly, even
when not using Harfbuzz (and to witness the additional
Harfbuzz magic when switching to it).

DrawTextString(): returns advance, so we can draw
subsegments of the text with the fallback font, instead
of only individual chars are previously.
This is needed to get correct shaping with Harfbuzz with
the fallback font.

Harbuzz, in both measureText() and DrawTextString():
Don't use filterChar() with the main font, as the
not-found chars might be found in the fallback font;
no need to replace them thar early.
Rework glyph/cluster/text walking and drawing to be more
generic: should work with "one char > multiple glyphs"
situations (we previously handled correctly only "multiple
chars merged into a single glyph", like ligatures), and
with RTL text.
Also accumulate not-found glyphs so we can draw them as
a single segment with the fallback font, with harfbuzz
(instead of with Freetype previously): this is needed
when using a latin font as the main font, and be able
to see nice arabic drawn with the fallback font.
Drop letterspacing when the detected script is cursive.

Harfbuzz light:
hbCalcCharWidth(): skip triplet when any of the 3 chars is
not found, as it would mess up the result (and cached values)
when some char is combined with mutiple rare diacritic marks
(eg. Hebrew). This could mess the drawing of the whole text.
Fallback to more robust Freetype measurement in such cases.

Repo: crengine



Repo: crengine