Optimisation: avoid re-init for EPUBs with embedded fonts
ImportEpubDocument() parses all .css in the EPUB looking for
@font-face before parsing the HTML files, but was registering
them only at the end, which needed a re-init to update node
styles with these fonts.
We now register these fonts before that and avoid this re-init.
As it also parses <HEAD><STYLE> while handling each HTML file,
still looking for @font-face, we keep track of any new one
found in head styles, and we only trigger the then neede
d re-init when we actually have found new ones (which should
be rare).
This should considerably speed up the initial loading of
EPUBs with embedded fonts.