Pagesplitting: account for lines' text direction
Mostly just to be able to put the in-page footnotes separator
on the right for RTL documents.
Switch LVRendPageInfo->type to ->flags to store additional info.
As text direction might change (albeit rarely) for elements on
a page, and as we don't have any global flag for the document
global direction (it may change per DocFragment), we just account
the nb of 'lines' and RTL 'lines', and flag the page (or footnotes
on that page) as RTL if we have > 50% of them RTL.

Repo: crengine



Repo: crengine