Text: allow wrap after '/' and '-'
With or without spaces around, as they are sometimes used
to separate words, and not only for math division.

Without spaces, have '/', just like '-', be candidate
for "deprecated wrap", just like most hyphens are.
With spaces, a wrap was set to be avoided previously
(as we were just thinking about math formulaes): removed
that rule.

This may avoid some spacing issues with line splitting
when such words are long.

Repo: crengine



Repo: crengine