Text: split measuring and word drawing by unicode script
When measuring with Harfbuzz, we should also split on Unicode script
change, even in a same bidi level (mixed hebrew and arabic in a single
text node should be handled as multiple segments, or Harfbuzz would
shape the whole segment with the script of the first kind of text it
meets, resulting in non-joined arabic if preceeded by hebrew).

Also remove First-Strong-Isolate unicode char from the list of RTL chars.
(Also fix wording of "bidi" as "bibi"...)

Repo: crengine


Repo: crengine