Text: look for hyphenation in more words if needed
When laying out lines and looking for wrap possibilities,
we use the last breakable space found. But we also try to
hyphenate the word that follows it to grab a bit of it.
In the case where there are more words after that last
breakable space (words separated by a non-breakable spaces,
or where our AVOID_WRAP_BEFORE/AFTER checks decided that
wrap is to be avoided at these spaces), we would only look
at hyphenating the last word of this series of words: if
that fails, we would not try to hyphenate previous words.
We know check hyphenation in all these words.

Sample test case: a french line ending with "pourqu'elle"
overflowing max line width (wrap at ' is avoided):
- if max line width ($) is met at "pourq$u'elle", "pour-qu"
  hyphenation would be found and used.
- if max line width is met at "pourqu'el$le", only "elle"
  would be checked, and no hyphenation found and we would wrap
  before "pourqu" ; "pour-qu" would not be considered at all.
  It will be now.

Repo: crengine


Repo: crengine