renderBlockElementEnhanced: compute baseline of block
Generic support for baseline of blocks. Will be needed to
support "display: inline-block" (boxes whose baseline is
the baseline of the last line box it contains) and
"inline-table" (boxes whose baseline is the baseline of
the first line box, or the bottom of the first table row).
- FlowState::addContentLine() takes an additional parameter,
  that we set when adding a line (ignored when adding paddings
  and margins).
- renderBlockElement() takes an optional *baseline so callers
  can get it additionally to the returned height.
- Also allows storing a baseline in RenderRectAccessor.

Repo: crengine


Repo: crengine