lvtextfm: dont adjust space after initial quotation mark/dash
If a paragraph starts with a quotation mark or dash, and is
followed by a space (classic or nbsp): don't have this space
width widened or shortened in the process of text justification.
This generalizes the existing code that was doing this only
for 4 kind of dashes: we do it for more opening marks, dashes,
bullets, and other ascii chars that could start a line.

Repo: crengine



Repo: crengine