lvtextfm: avoid spurious spaces when hyphenating (bis)
Since fa0529f and possibly looking for hyphenation on more
than one word word, we may set hyph flags at multiple positions,
and have non last words larger than they should, causing the
appearance of a space.
So, kill any hyphen flag we won't be using.
This reverts 57ed741, a previous attempt which was fine in LTR
text, but killed the hyphen in BiDi text (where the hyphenated
word might not be the (visual) last on line).