Page splitting: split single column tables by lines instead of row
If table is a single column, we can add to main context
the lines of each cell, instead of full table rows, which
will avoid cell lines to possibly be cut in the middle.
So, when table is single column, we transfer the lines from
the individual LVRendPageContext used when rendering each
cell to the parent LVRendPageContext (which works recursively)
up to the main LVRendPageContext that will do the book pages
splitting.
Adapt a bit LVRendPageContext and FlowState to distinguish
hasPageList vs wantsLine.

Repo: crengine



Repo: crengine