toStringV2(): output [1] when first but not single
toStringV1() would output "/div/" when the element
is the single DIV in its parent children, but would
output "/div[1]/" when parent has more than one DIV
and element is the first one of them, which is nice.
toStringV2() was outputing "div" in both cases.
This has toStringV2() do as toStringV1().
(Note that createXPointerV2() would find the correct
node with both kind of input.)

Repo: crengine



Repo: crengine