FB2 footnotes: fix spurious newline ("display: run-in")
"display: run-in" (not supported by all browsers and possibly
deprecated) support might have been added only as a way to solve
a rendering issue with FB2 footnotes, where the footnote number
is in a block element, and the footnote text in another block
element. fb2.css sets the first block to be "display: run-in"
as an attempt to bring them both in an erm_final element so
they can be rendered on the same line.
This fixes the fact this was not happening when there is
a text node (a single "\n" in the HTML) between these two
nodes.

Repo: crengine



Repo: crengine