Fonts: allow providing and using multiple fallback fonts
As our use of the fallback font was a recursive calls, we can
provide chained fallback fonts.
As one of the fallback font could be the main font, we need to
use first getFallbackFont() (to get the first fallback font, even
on a font among these fallback fonts), and getNextFallbackFont()
to get the chained fallback font.
Caveat: we might try rendering/drawing with the main font twice
if it is among the fallback fonts (but avoiding that wouldn't
make this change as simple).

Repo: crengine



Repo: crengine