Strut confinning: deal with images earlier
Rework cedbe825 for images, by checking and adjusting
only once in measureText(), instead of using a wrong
trick in ProcessParagraph() (which could mess widths
in adjusted image does not fit on the line).
This ends up also not having to recompute the image
dimensions in AddLine().

Repo: crengine



Repo: crengine