CSS: support for pseudo elements ::before & ::after
Handle parsing of '::before', '::after' (CSS3), ':before'
and ':after (CSS2) in selectors.
Properly check if they should be generated or not, and if
yes, insert a new internal element in the DOM: pseudoElem.

Handle needed added CSS property: 'content:'. Parse original
values and store a pre-computed string in style->content,
ready to be used to get the final generated content for
a node. Supports string, attributes, open/close-quote.

Replaces <Q> specific handling in the code with:
  q::before { content: open-quote; }
  q::after  { content: close-quote; }

Repo: crengine



Repo: crengine