Text: skip formatting optimisations when inlineBoxes
We need to go thru alignLine() when inlineBoxes are involved.
Since 6c3f6e4e, when leaving a small book (that would not
generate a cache file) on a page starting with some
inline-block, when re-opening it and jumping to that
xpointer (before the paragraph is drawn), it would
find it at top left (0,0) of paragraph and could go
to a previous page that contains the paragraph start.