styleToTextFmtFlags(): adds is_block parameter
Checking (style->display > css_d_inline) to decide
the node is a block node might not always be right.
We should check the rendering method, which might be
different than the specified CSS "display:" (it will
be with ruby). Even more flexible to have the calling
code tells it.

Repo: crengine



Repo: crengine