Adds support for "display: ruby" and <ruby> elements
Adds support for "display: ruby", which when set on a ruby
element, will wrap its content in a few inlineBox and rubyBox
elements to build a inline-table like DOM fragment so we can
render it with our standard table rendering code.
The standard table rendering code will just swap the first
2 lines so the base makes the 2nd row and the first annotation
the 1st row.
Also, we limit interline up-scaling for ruby to allow using
it to make room when needed with ruby.

Repo: crengine



Repo: crengine