Disable hasCacheFile()/setBoxingWishedButPreventedByCache()
With previous commit, that seems to solve the crashes we
used to get when boxing/removing/moving nodes around when
there is a cache file, we might no more need these
hasCacheFile() checks and _boxingWishedButPreventedByCache
flag.
(Not fully removing them for now, just in case we notice
some other issue and decide this might still be prefered.)

We still keep updating _nodeDisplayStyleHash when the CSS
display property changes for a node, as a previous boxing
might have removed some whitespace only text node when a
node was block, that might have some value when the node
becomes inline. So, isBuiltDomStale() will continue to be
triggered and propose a full reload, even if a proper
boxing is now being made - and the rendering could be
totally fine.

Repo: crengine



Repo: crengine