globalHash: exclude some settings on initial loading
If not yet rendered (initial loading with XML parsing), we can
ignore some global flags that have not yet produced any effect,
so they can possibly be updated between loading and rendering
without trigerring a drop of all the styles and rend methods
set up in the XML loading phase. This is mostly only needed
for TextLangMan::getHash(), as the lang can be set by frontend
code after the loading phase, once the book language is known
from its metadata, before the rendering that will use the
language set.
With KOReader, this should speed up a book first load where
its language is not yet known, and is different from the
default or fallback typography language.

Repo: crengine


Repo: crengine