CSS: optimize pseudoclass nth-child() and friends
The getUnboxedSibling calls are expensive, and we
can cache the resulting index of a node in its parent
children collection as it won't change.
We can cache these values in a node's RenderRectAccessor,
as it's not yet used at this point.
(On some books with a few dozens of p:nth-child(N)
rules, it can reduce the loading and rendering from
a few minutes to only a few seconds.)

Repo: crengine


Repo: crengine