Avoid clang warnings: use of memory after it is freed
clang-tidy does not really handle well our reference
counting classes instances like LVRef.
The code was working fine and does not need any fix,
but to silence these warnings, instead of fetching
the css_style_rec_t or LVFont, we keep using the LVRef
(that wrap these) we get from a node. This should be
equivalent to the previous code (with may be an added
indirection to resolve more often).

Repo: crengine



Repo: crengine