FB2 footnotes: fix spacing between number and text
With "display: run-in", only used for FB2 footnotes, the
footnote numbers and the text may or may not end or start
with a space.
Previously, 2 nbsp were added so we could detect them to
ensure the width would not vary with justification - but
this resulted in all these spaces not collapsing and a
final width variable from one document to another.
Now, we just add a normal space, and a flag, so they all
collapse into one that will get a non-variable width.

Also add a space after "display: run-in" content so that
FB2 footnotes get a space (even when there is none in
the source) when providing it to MuPDF.

Repo: crengine



Repo: crengine