More generic cleanup of RenderRect used for CSS caching
(Follow up to 4d307aea and 58e869f0.)
We should always clean it up after the initial XML
loading phase (otherwise, even if not needed for
this initial phase, it could spill over the next
re-rendering and mess the 2nd rendering).
Also fix ldomDataStorageManager::setRendRectData()
possible crash; getRenderRectData() is doing that
right, but setRendRectData() was possibly not
allocating the chunk we're about to use.