(Upstream) FB2: fix coverpage drawing in scroll mode
Happened only in enhanced rendering mode, caused by
forgetting to use the 'y' provided to RenderBlockElement().
As we're now using it, don't provide a float's container
padding_top as this 'y' when rendering the float, as it has
already been given to and accounted by the main FlowState.

Repo: crengine


Repo: crengine