FB2 footnotes: only merge run-in when next is erm_final
The content following the 'display:run-in' footnote number
is usually a P rendered erm_final, and this merging of them
into a container erm_final works quite well.
But when the content is not a P but some more complex
erm_block structure (like poem>stanza>v or cite>p),
ensuring this mergins would make that structure all
erm_inline, losing the formatting. So don't do that,
and let a blank line after the footnote number happen,
which is less worse than losing formatting.

Repo: crengine



Repo: crengine