(Upstream) Rename lChar16/lString16 to lChar32/lString32
As all these were actually stored in 32bits on the
platforms we support.
Upstream needs this change to be able to re-introduce
real lChar16/lString16 for use with Windows API. This
additional change was not picked, as not needed and so
we don't accidentally use these crippled 16bits types.

Massive nearly brainless renames of:
- lChar16 to lChar32
- lString16 to lString32
- unused lStringBuf16 to lStringBuf32
- wchar_t to lChar32 or char32_t
- cs16() to cs32()
- L" to U", L' to U', L to U in some macros
- some other 16 to 32 in variable names

Repo: crengine


Repo: crengine