In-page footnotes: always push vertical margin before
When the footnote container has some padding-top,
any previous accumulated vertical margin was pushed,
and the padding-top added, before entering the
footnote - so, collapsing margins and padding-top
are not included in the in-page footnote.
Do this also when it has 0 padding-top, for consistency
and to avoid any previous vertical margin to be in the
in-page footnotes (which was often happening with the
first footnote, getting the bottom margin of its
preceeding H1 or H2 header).

Repo: crengine


Repo: crengine