(Upstream) CacheFileHeader: zero the alignment padding for reproducibility
The CacheFileItem struct within the CacheFileHeader was
being aligned by the compiler to an 8 byte boundary creating
a 4 byte whole in the header that was not being zeroed and
this was leading to unreproducible content in the cache file.
Add the padding explicitly and zero it.

Repo: crengine


Repo: crengine