LVFontCache::find(): give more weight to first fonts in list
When there are multiple font names in a list (ie, with
CSS 'font-family: "Font1", "Font2"', multiple fonts may
get the same higher score. We keep the first best match
in the loop, but at the end, we prefer an instantiated
font to a not yet instantiated one with the same match.
Depending on various things, it may not be the first one.
So, provide an additional weight to font by their order
in the list.