CSS: add units 'ch' (just like 'ex')
Represents the width, or more precisely the advance measure,
of the glyph "0" (zero) in the element's font.
In the cases where it is impractical to determine, it must
be assumed to be 0.5em wide by 1em tall.
So, let's handle it just as 'ex', assuming 0.5em.

Repo: crengine



Repo: crengine