sunxi: Bypass dubious driver optimizations to ensure flashing actually flashes
I'm getting retro einkfb vibes...

The memcmp is new in the Sage kernel (we'll see if it hits the Elipsa).

On the upside, at first blush,
the other disp tweaks make mixing layouts much more viable,
and probably make the whole thing behave closer to mxcfb, which
is good ;).

Repo: base


Repo: base