Hanging punctuation: tweak hanging ratios again
Hanging ratio was a bit too heavy for guillemets.
Make it lower, and dependant on the side (instead
of the presence of a space alongside).
Pass LTR/RTL status of the paragraph so we can
know whether left or right side is start or end
of line, and return ratios accordingly.

Repo: crengine



Repo: crengine