Fix sizing of float/inline-block image containers
The floatBox or inlineBox container of an image could
be larger than then drawn image (ie. when the image is
itself larger than the screen size) as images are
later resized to fit the screen. Account for that
at rendering time, so we don't get empty space after
the image in such containers.

Repo: crengine


Repo: crengine