Text: correct glyph overlap at text node boundaries
This is usually known as "italic correction", but we can
do it more generically for all text nodes, as it can help
too with regular fonts with different line-height or
vertical-align.
We don't use font and glyph metrics like side bearings, as
this is quite limited: we do it visually by drawing both left
and right words in a new LVHorizontalOverlapMeasurementDrawBuf,
and getting the collision distance from the real glyphs used
by the final drawing.

Repo: crengine


Repo: crengine