Hyphenation: handle combining diacritics better
Previously, hyphenation could happen between a letter
and its diacritic, leaving the diacritic dangling on
next line.
So, be sure we correctly skip diacritics/modifiers:
remove them from the word to be matched, and when
applying flag to the input string, be sure to apply
any flag after any diacritic.
As some pattern files (Latin, Greek) do contains
some diacritics, check which are present at load time,
and don't skip these ones (and only these ones): this
should ensure existing patterns will still match what
they aimed at, but it expects the patterns to handle
all cases with these diacritics (which may or may not
be the case for Latin and Greek).