Hyphenation: rework AlgoHyph::hyphenate()
The code was a bit convoluted and looks like it aimed
at supporting multiple words in the input string, which
does not happen the way we use it. So, skip this confusing
outer loop.
State how and what it is supposed to do (from what could
be gathered from the existing code, which feels logical).
Fix left/right hyphen min ensurance, and properly handle
diacritics/modifiers as done in previous commit for TexHyph.

Repo: crengine



Repo: crengine