CJK: expand last line as previous justified line
The last line of a justified paragraph is left aligned and
wouldn't get any added spacing, which would make it look
more condensed that the justified line above it.
So, when both previous to last and last line starts with
2 CJK chars, add to the last line the same spacing added
to the above line so CJK chars looks vertically aligned.

Repo: crengine


Repo: crengine