SVG: enhanced SVG support with LunaSVG extended
Add support for SVG image rendering with the help
of LunaSVG, patched/extended to use crengine to
handle <text>, <tspan> and <image>.
Add LVImageSource:IsScalable() and allow DrawBuf
to get requested the target image size so
LVSvgImageSource::Decode() can have LunaSVG draw
to this size so no bitmap scaling is needed.
Add LVSvgImageSource::Render() so frontend can
request a SVG image to be rendered (by LunaSVG)
as a bitmap of the requested size (not implemented
for other image types).

Repo: crengine


Repo: crengine