Handle the <svg> element as an SVG image
Properly parse the <svg> subtree, keeping elements
and attributes case.
Allow serializing it back to a string, so it can be
fed to LunaSVG. While doing that, add some bits from
the crengine DOM computed styles so any HTML-level
stylesheets that styles the <svg> children may not
be ignored.
Everywhere we handle images, accept <svg> and use
writeSVGNode() to create an image content stream.
Try to handle correctly SVG wrappers of cover images
that uses width/height in %.
Don't handle <svg> as image when in legacy rendering
mode: this might be useful to get text inside SVG as
plain crengine text, and allow search in it.