xtext: properly support soft-hyphens
If soft-hyphens are present in the input text, they are
considered (by libunibreak) as candidates for breaking
a line, but no real hyphen would be shown at end of
line when the line is rendered.
So, be sure we properly draw a real hyphen when shaping
a line ending with a soft-hyphen.
(Refactor the ellipsis substitution code so we can reuse
it with hyphen substitution.)

Repo: base



Repo: base