CSS: cache parse results (koreader/crengine#493)
Stylesheets included in each DocFragment are parsed individually,
even if they point to the same file. Depending on the number of
CSS files, their size, and number of DocFragments, CSS parsing
accounts for several percentage of a full LVDocView::Render() time,
or even more in extreme cases. By introducing a parsing cache,
the time spent in LVImportStylesheetParser::Parse is now negligible.