EPUB: support obfuscated embedded fonts
There is code to demangle fonts obfuscated by the Adobe
algorithm, but it has never been triggered for 10+ years
because, although META-INF/encryption.xml was parsed,
its items where not stored and the list of encrypted
items stayed empty!
So, properly handle that, and also add support for the
alternative algorithm from the IDPF/EPUB specs (which
needs us to compute a sha1 of the book identifier, so
add a sha1digest() function).
The "Reading of DRM protected books is unsupported"
message was also never shown: show it now when we
notice all items are encrypted with an unsupported
algorithm.