mxcfb: Implement fenced updates modes ([partial] & [ui]) (koreader/koreader-base#1576)
There's no per-update flag on mxcfb (instead there are global no-merge
UPDATE_SCHEME modes). Since we don't want to mess with the global mode,
we fake the fencing via explicit wait_for ioctl usage.
To account for devices with broken wait_for ioctl (since those happen to
be the target for testing this implementation), we force a much larger
sleep than usual for the pre- fence (250 instead of 2.5ms).