LVStream: disable CRC checks for ZIP streams
Rational: in most case, CRC errors are completely ignored, making
calculating/checking it a waste of CPU. In the rare instances it's
not ignored, the behavior is far from ideal (e.g. if a FB3 file
metadata is corrupted, Koreader will print an error message and
flat-out just quit!).

Repo: crengine



Repo: crengine