MuPDF: fix build (koreader/koreader-base#1586)
Patch the makefile link command so the build process does not try to link with system libraries (e.g. `-lpeg`), even when building the generator helpers (like `cnamedump.exe`). Since mupdf's build system is only in charge of building a static library, we can just drop the `$(LIBS)` part of the link command.